+++ zanshinsdk/test_client.py
@@ -434,76 +434,6 @@ class TestClient(unittest.TestCase):
         )
 
     ###################################################
-    # Organization Member
-    ###################################################
-
-    def test_iter_organization_members(self):
-        organization_id = "822f4225-43e9-4922-b6b8-8b0620bdb1e3"
-
-        try:
-            next(self.sdk.iter_organization_members(organization_id))
-        except StopIteration:
-            pass
-
-        self.sdk._request.assert_called_once_with(
-            "GET", f"/organizations/{organization_id}/members"
-        )
-
-    def test_get_organization_members(self):
-        organization_id = "822f4225-43e9-4922-b6b8-8b0620bdb1e3"
-        member_id = "e22f4225-43e9-4922-b6b8-8b0620bdb110"
-
-        self.sdk.get_organization_member(organization_id, member_id)
-
-        self.sdk._request.assert_called_once_with(
-            "GET", f"/organizations/{organization_id}/members/{member_id}"
-        )
-
-    def test_update_organization_members(self):
-        organization_id = "822f4225-43e9-4922-b6b8-8b0620bdb1e3"
-        member_id = "e22f4225-43e9-4922-b6b8-8b0620bdb110"
-        role = [zanshinsdk.Roles.ADMIN]
-
-        self.sdk.update_organization_member(organization_id, member_id, role)
-
-        self.sdk._request.assert_called_once_with(
-            "PUT",
-            f"/organizations/{organization_id}/members/{member_id}",
-            body={"roles": role},
-        )
-
-    def test_delete_organization_members(self):
-        organization_id = "822f4225-43e9-4922-b6b8-8b0620bdb1e3"
-        member_id = "e22f4225-43e9-4922-b6b8-8b0620bdb110"
-
-        self.sdk.delete_organization_member(organization_id, member_id)
-
-        self.sdk._request.assert_called_once_with(
-            "DELETE", f"/organizations/{organization_id}/members/{member_id}"
-        )
-
-    def test_reset_organization_member_mfa(self):
-        organization_id = "822f4225-43e9-4922-b6b8-8b0620bdb1e3"
-        member_id = "e22f4225-43e9-4922-b6b8-8b0620bdb110"
-
-        self.sdk.reset_organization_member_mfa(organization_id, member_id)
-
-        self.sdk._request.assert_called_once_with(
-            "POST", f"/organizations/{organization_id}/members/{member_id}/mfa/reset"
-        )
-
-    def test_reset_delete_organization_password(self):
-        organization_id = "822f4225-43e9-4922-b6b8-8b0620bdb1e3"
-        member_id = "e22f4225-43e9-4922-b6b8-8b0620bdb110"
-
-        self.sdk.reset_delete_organization_password(organization_id, member_id)
-
-        self.sdk._request.assert_called_once_with(
-            "POST",
-            f"/organizations/{organization_id}/members/{member_id}/password/reset",
-        )
-
-    ###################################################
     # Organization Member Invite
     ###################################################
